+++ HVNixieDriver.h
@@ -20,7 +20,7 @@ protected:
 	int LEpin;
 
 	virtual uint8_t getSPIMode() = 0;
-	virtual uint32_t NIXIE_DRIVER_ISR_FLAG getPins(byte mask) = 0;
+	virtual uint64_t NIXIE_DRIVER_ISR_FLAG getPins(byte mask) = 0;
 	virtual uint32_t NIXIE_DRIVER_ISR_FLAG getPin(uint32_t digit) = 0;
 	virtual uint32_t NIXIE_DRIVER_ISR_FLAG getMultiplexPins();
 	virtual uint32_t NIXIE_DRIVER_ISR_FLAG convertPolarity(uint32_t pins) = 0;
